+++ Dimensia/Dimensia/src/render/GuiTitle.java
@@ -18,7 +18,7 @@ public class GuiTitle extends GuiComponent
 	 * @param width the width of the button, as a percent of the screen, from 0.0F to 1.0F
 	 * @param height the height of the button, as a percent of the screen, from 0.0F to 1.0F
 	 */
-	public GuiTitle(String value, float x, float y, float width, float height)
+	public GuiTitle(String value, double x, double y, double width, double height)
 	{
 		super();
 		this.x = x;
@@ -40,10 +40,10 @@ public class GuiTitle extends GuiComponent
 //		renderTexture.bind();
 //		t.startDrawingQuads();
 		
-		float width = Display.getWidth() * this.width * 0.5F;
-		float height = Display.getHeight() * this.height * 0.5F;
-		float x = Display.getWidth() * this.x * 0.5F;
-		float y = Display.getHeight() * this.y * 0.5F;
+		double width = Display.getWidth() * this.width * 0.5F;
+		double height = Display.getHeight() * this.height * 0.5F;
+		double x = Display.getWidth() * this.x * 0.5F;
+		double y = Display.getHeight() * this.y * 0.5F;
 		
 		//if(width < (trueTypeFont.getWidth(values[buttonIndex]) / 2))
 		//{
@@ -58,8 +58,8 @@ public class GuiTitle extends GuiComponent
 //		
 		
 		//Text:
-		float xOffset = x + width/2 * 0.95f;//trueTypeFont.getWidth(values[buttonIndex]) / 2;
-		float yOffset = y + height - (height - trueTypeFont.getHeight(value)) / 2;
+		float xOffset = (float) (x + width/2 * 0.95f);//trueTypeFont.getWidth(values[buttonIndex]) / 2;
+		float yOffset = (float) (y + height - (height - trueTypeFont.getHeight(value)) / 2);
 			
 		trueTypeFont.drawString(xOffset, yOffset, value, 0.7f, -1, TrueTypeFont.ALIGN_CENTER); //Render the Text	
 	}
